more cleanup
this should give also give us a speedup. there was never any need for
now and then to be floats

+++ losthorizons/baseapplication.h
@@ -31,7 +31,7 @@ private:
 	void getBits();
 
 	//for delta time calculation
-	f32 then;
+	u32 then;
 };
 
 #endif
